MNT: refactor to avoid importing numpy in utils
