+++ tiled/adapters/sequence.py
@@ -1,4 +1,6 @@
 import builtins
+import math
+import warnings
 from abc import abstractmethod
 from pathlib import Path
 from typing import Any, List, Optional, Tuple, Union
@@ -12,7 +14,47 @@ from ..structures.core import Spec
 from ..utils import path_from_uri
 from .protocols import AccessPolicy
 from .type_alliases import JSON, NDSlice
-from .utils import force_reshape
+
+
+def force_reshape(arr: np.array, desired_shape: Tuple[int, ...]) -> np.array:
+    """Reshape a numpy array to match the desited shape, if possible.
+
+    Parameters
+    ----------
+
+    arr : np.array
+        The original ND array to be reshaped
+    desired_shape : Tuple[int, ...]
+        The desired shape of the resulting array
+
+    Returns
+    -------
+
+    A view of the original array
+    """
+
+    if arr.shape == desired_shape:
+        # Nothing to do here
+        return arr
+
+    if arr.size == math.prod(desired_shape):
+        if len(arr.shape) != len(desired_shape):
+            # Missing or extra singleton dimensions
+            warnings.warn(
+                f"Forcefully reshaping {arr.shape} to {desired_shape}",
+                category=RuntimeWarning,
+            )
+            return arr.reshape(desired_shape)
+        else:
+            # Some dimensions might be swapped or completely wrong
+            # TODO: needs to be treated more carefully
+            pass
+
+    warnings.warn(
+        f"Can not reshape array of {arr.shape} to match {desired_shape}; proceeding without changes",
+        category=RuntimeWarning,
+    )
+    return arr
 
 
 class FileSequenceAdapter:
